// src/app/models/Movie/filtreMovie.model.ts
export class FiltreMovie{

    title: string;
    realisateur: string;
    scenariste: string;
    producteur: string;
    acteur: string[];
    sortie: Date;
    categories: string[];
    isPret: boolean;
    nameOfPret: string;
    isTVShow: boolean;
    saisons: string;


    constructor(){
    }

}